requirements.txt
Babel==2.6.0
certifi==2018.11.29
Click==7.0
Django==2.1.5
django-crispy-forms==1.7.2
django-phonenumber-field==2.2.0
dlib==19.16.0
face-recognition==1.2.3
face-recognition-models==0.3.0
numpy==1.16.1
opencv-python==4.0.0.21
phonenumbers==8.10.5
Pillow==5.4.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
pytz==2018.9
